# Optimize landing page performance to achieve 95%+ Lighthouse scores