--- a/app/layout.tsx
+++ b/app/layout.tsx
@@ -6,7 +6,12 @@ import Script from 'next/script';
 import ClientProviders from '../components/ClientProviders';
 import type { Metadata } from 'next';
 
-const inter = Inter({ subsets: ['latin'] });
+const inter = Inter({ 
+  subsets: ['latin'],
+  display: 'swap',
+  preload: true,
+  fallback: ['system-ui', 'arial']
+});
 
 export const metadata: Metadata = {
   title: 'Easy Close Properties - We Buy Houses Fast',
@@ -21,13 +26,51 @@ export default function RootLayout({
   return (
     <html lang="en">
       <head>
+        {/* Preconnect to critical third-party origins */}
+        <link rel="preconnect" href="https://www.google.com" />
+        <link rel="preconnect" href="https://www.gstatic.com" crossOrigin="anonymous" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
+        <link rel="preconnect" href="https://www.googletagmanager.com" />
+        <link rel="dns-prefetch" href="https://maps.googleapis.com" />
+        <link rel="dns-prefetch" href="https://maps.gstatic.com" />
+        
+        {/* Critical CSS for above-the-fold content */}
+        <style dangerouslySetInnerHTML={{
+          __html: `
+            /* Critical CSS for initial render */
+            * { margin: 0; padding: 0; box-sizing: border-box; }
+            body { font-family: system-ui, -apple-system, sans-serif; line-height: 1.5; }
+            .min-h-screen { min-height: 100vh; }
+            .bg-gradient-to-b { background: linear-gradient(to bottom, #f9fafb, #ffffff); }
+            .bg-gradient-to-br { background: linear-gradient(to bottom right, #002767, #bd0b31); }
+            .text-white { color: white; }
+            .text-center { text-align: center; }
+            .mx-auto { margin-left: auto; margin-right: auto; }
+            .max-w-6xl { max-width: 72rem; }
+            .max-w-3xl { max-width: 48rem; }
+            .pt-20 { padding-top: 5rem; }
+            .pb-16 { padding-bottom: 4rem; }
+            .px-4 { padding-left: 1rem; padding-right: 1rem; }
+            .mb-6 { margin-bottom: 1.5rem; }
+            .mb-8 { margin-bottom: 2rem; }
+            .text-4xl { font-size: 2.25rem; }
+            .font-bold { font-weight: 700; }
+            .text-xl { font-size: 1.25rem; }
+            @media (min-width: 768px) {
+              .md\:text-5xl { font-size: 3rem; }
+            }
+            @media (min-width: 1024px) {
+              .lg\:text-6xl { font-size: 3.75rem; }
+            }
+          `
+        }} />
         <Script
-          strategy="afterInteractive"
+          strategy="lazyOnload"
           src="https://www.googletagmanager.com/gtag/js?id=AW-17109864760"
         />
         <Script
           id="gtag-init"
-          strategy="afterInteractive"
+          strategy="lazyOnload"
           dangerouslySetInnerHTML={{
             __html: `
               window.dataLayer = window.dataLayer || [];
@@ -58,22 +101,6 @@ export default function RootLayout({
           </main>
           <Footer />
         </ClientProviders>
-        <Script
-          src={`https://maps.googleapis.com/maps/api/js?key=${process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY}&libraries=places&callback=initGoogleMaps`}
-          strategy="lazyOnload"
-        />
-        <Script
-          id="google-maps-init"
-          strategy="lazyOnload"
-          dangerouslySetInnerHTML={{
-            __html: `
-              window.initGoogleMaps = function() {
-                console.log('Google Maps initialized');
-                window.dispatchEvent(new Event('google-maps-loaded'));
-              };
-            `,
-          }}
-        />
       </body>
     </html>
   );