… public/ECPtransparent.webp
… public/Easy Close Properties Logo.webp
… public/Easy Close Properties Transparent.webp
… public/favicon.ico
… reports/lh-desktop-after.json
… reports/lh-desktop-baseline.json
… reports/lh-mobile-after.json
… reports/lh-mobile-baseline.json