## Summary
- Achieved 98% mobile and 100% desktop performance scores (up from 71% and 78%)
- Implemented comprehensive performance optimizations without changing functionality
- Fixed accessibility issues and added security headers

## Performance Improvements
- **Mobile Performance**: 71% → 98% ✅
- **Desktop Performance**: 78% → 100% ✅
- **LCP**: Improved by 56% (3.4s → ~1.5s)
- **TBT**: Reduced by 89% (924ms → ~100ms)

## Key Changes
### 🚀 Performance
- Lazy load Google Maps API - only loads when user interacts with address field (saves 281KB)
- Inline critical CSS for above-the-fold content
- Add aggressive caching headers for static assets
- Convert images to WebP format with 85% quality

### 🔒 Security & Best Practices
- Add comprehensive security headers (CSP, X-Frame-Options, etc.)
- Implement proper cache control policies
- Add favicon to resolve 404 errors

### ♿ Accessibility
- Fix color contrast issues in footer (removed opacity from white text)
- Ensure all text meets WCAG AA standards (4.5:1 minimum contrast)

### 📱 Mobile UX
- Fix header button overlap on mobile devices
- Make logo and buttons responsive across breakpoints
- Optimize text display for small screens

## Test Results
All Lighthouse scores now exceed 95% target except Best Practices at 93% (due to necessary third-party scripts).

## Files Changed
- Core optimizations in `next.config.js`, `app/layout.tsx`
- New lazy loading hook: `hooks/useGoogleMaps.ts`
- Updated components for accessibility and mobile responsiveness
- Added WebP versions of images and performance audit reports